# backend/docs/swagger.yaml
definitions:
...
  credential.CategoryCredential:
    properties:
      name:
        type: string
      password:
        type: string
      private:
        type: boolean
    type: object
  credential.HttpsSettingCredential:
    properties:
      redirect:
        type: boolean
    type: object
...
  domain.Article:
    properties:
      author:
        type: string
      category:
        type: string
      content:
        type: string
      copyright:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      hidden:
        type: boolean
      id:
        type: integer
      lastVisitedTime:
        type: string
      password:
        type: string
      pathname:
        type: string
      private:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      top:
        type: integer
      updatedAt:
        type: string
      viewer:
        type: integer
      visited:
        type: integer
    type: object
...
  domain.Category:
    properties:
      id:
        type: integer
      meta: {}
      name:
        type: string
      password:
        type: string
      private:
        type: boolean
      type:
        $ref: '#/definitions/domain.CategoryType'
    type: object
  domain.CategoryType:
    enum:
    - category
    - column
    type: string
    x-enum-varnames:
    - CategoryCategoryType
    - ColumnCategoryType
...
  domain.HttpSetting:
    properties:
      redirect:
        type: boolean
    type: object
...
  domain.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      nickname:
        type: string
      password:
        type: string
      permissions:
        items:
          type: string
        type: array
      salt:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/caddy/ask:
    get:
      consumes:
      - application/json
      description: acquire https settings
      parameters:
      - description: domain
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: acquire https settings
      tags:
      - Caddy
...
  /api/admin/caddy/https:
    get:
      consumes:
      - application/json
      description: acquire https settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpSetting'
      summary: acquire https settings
      tags:
      - Caddy
    put:
      consumes:
      - application/json
      description: update https settings
      parameters:
      - description: httpsSettingCredential
        in: body
        name: httpsSettingCredential
        required: true
        schema:
          $ref: '#/definitions/credential.HttpsSettingCredential'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update https settings
      tags:
      - Caddy
...
  /api/admin/category:
    post:
      consumes:
      - application/json
      description: create category
      parameters:
      - description: credential
        in: body
        name: credential
        required: true
        schema:
          $ref: '#/definitions/credential.CategoryCredential'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: create category
      tags:
      - Category
  /api/admin/category/:name:
    delete:
      consumes:
      - application/json
      description: delete category by category name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: delete category by category name
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: get articles by category name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Article'
            type: array
      summary: get articles by category name
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: delete category by category name
      parameters:
      - description: credential
        in: body
        name: credential
        required: true
        schema:
          $ref: '#/definitions/domain.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update category by category name
      tags:
      - Category
...
  /api/admin/collaborator:
    get:
      consumes:
      - application/json
      description: get all collaborator
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
      summary: get all collaborator
      tags:
      - Collaborator
    post:
      consumes:
      - application/json
      description: create collaborator
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: create collaborator
      tags:
      - Collaborator
    put:
      consumes:
      - application/json
      description: update collaborator
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update collaborator
      tags:
      - Collaborator
...
swagger: "2.0"